+++ src/Core/Py.jl
@@ -168,16 +168,23 @@ function Base.show(io::IO, ::MIME"text/plain", o::Py)
     end
     hasprefix = (get(io, :typeinfo, Any) != Py)::Bool
     compact = get(io, :compact, false)::Bool
+    limit = get(io, :limit, true)::Bool
+    if compact
+        # compact should output a single line, which we force by replacing newline
+        # characters with spaces
+        str = replace(str, "\n" => " ")
+    end
     multiline = '\n' in str
-    prefix =
-        hasprefix ?
-        compact ? "Py:$(multiline ? '\n' : ' ')" : "Python:$(multiline ? '\n' : ' ')" : ""
+    prefix = !hasprefix ? "" : compact ? "Py: " : multiline ? "Python:\n" : "Python: "
     print(io, prefix)
     h, w = displaysize(io)
-    if get(io, :limit, true)
+    if limit
+        # limit: fit the printed text into the display size
         h, w = displaysize(io)
         h = max(h - 3, 5) # use 3 fewer lines to allow for the prompt, but always allow at least 5 lines
         if multiline
+            # multiline: we truncate each line to the width of the screen, and skip
+            #   middle lines if there are too many for the height
             h -= 1 # for the prefix
             lines = split(str, '\n')
             function printlines(io, lines, w)
@@ -218,7 +225,25 @@ function Base.show(io::IO, ::MIME"text/plain", o::Py)
                 println(io)
                 printlines(io, lines[end-h1+1:end], w)
             end
+        elseif compact
+            # compact: we print up to one screen width, skipping characters in the
+            #    middle if the string is too long for the width
+            maxlen = w - length(prefix)
+            if length(str) ≤ maxlen
+                print(io, str)
+            else
+                gap = " ... "
+                gaplen = length(gap)
+                w0 = cld(maxlen - gaplen, 2)
+                i0 = nextind(str, 1, w0 - 1)
+                i1 = prevind(str, ncodeunits(str), maxlen - gaplen - w0 - 1)
+                print(io, str[begin:i0])
+                printstyled(io, gap, color = :light_black)
+                print(io, str[i1:end])
+            end
         else
+            # single-line: we print up to one screenfull, skipping characters in the
+            #    middle if the string is too long. We skip a whole line of characters.
             maxlen = h * w - length(prefix)
             if length(str) ≤ maxlen
                 print(io, str)
@@ -257,25 +282,29 @@ Base.setproperty!(x::Py, k::Symbol, v) = pysetattr(x, string(k), v)
 Base.setproperty!(x::Py, k::String, v) = pysetattr(x, k, v)
 
 function Base.propertynames(x::Py, private::Bool = false)
-    # this follows the logic of rlcompleter.py
-    function classmembers(c)
-        r = pydir(c)
-        if pyhasattr(c, "__bases__")
-            for b in c.__bases__
-                r = pyiadd(r, classmembers(b))
+    properties = C.on_main_thread() do
+        # this follows the logic of rlcompleter.py
+        function classmembers(c)
+            r = pydir(c)
+            if pyhasattr(c, "__bases__")
+                for b in c.__bases__
+                    r = pyiadd(r, classmembers(b))
+                end
             end
+            return r
         end
-        return r
-    end
-    words = pyset(pydir(x))
-    words.discard("__builtins__")
-    if pyhasattr(x, "__class__")
-        words.add("__class__")
-        words.update(classmembers(x.__class__))
-    end
-    words = map(pystr_asstring, words)
+
+        words = pyset(pydir(x::Py))
+        words.discard("__builtins__")
+        if pyhasattr(x, "__class__")
+            words.add("__class__")
+            words.update(classmembers(x.__class__))
+        end
+        map(pystr_asstring, words)
+    end::Vector{String} # explicit type since on_main_thread() is type-unstable
+
     # private || filter!(w->!startswith(w, "_"), words)
-    map(Symbol, words)
+    map(Symbol, properties)
 end
 
 Base.Bool(x::Py) = pytruth(x)
@@ -334,31 +363,22 @@ Base.broadcastable(x::Py) = Ref(x)
 (f::Py)(args...; kwargs...) = pycall(f, args...; kwargs...)
 
 # comparisons
-Base.:(==)(x::Py, y::Py) = pyeq(x, y)
-Base.:(!=)(x::Py, y::Py) = pyne(x, y)
-Base.:(<=)(x::Py, y::Py) = pyle(x, y)
-Base.:(<)(x::Py, y::Py) = pylt(x, y)
-Base.:(>=)(x::Py, y::Py) = pyge(x, y)
-Base.:(>)(x::Py, y::Py) = pygt(x, y)
+Base.:(==)(x::Py, y::Py) = pyeq(Bool, x, y)
+Base.:(<=)(x::Py, y::Py) = pyle(Bool, x, y)
+Base.:(<)(x::Py, y::Py) = pylt(Bool, x, y)
 Base.isless(x::Py, y::Py) = pylt(Bool, x, y)
 Base.isequal(x::Py, y::Py) = pyeq(Bool, x, y)
 
 # we also allow comparison with numbers
-Base.:(==)(x::Py, y::Number) = pyeq(x, y)
-Base.:(!=)(x::Py, y::Number) = pyne(x, y)
-Base.:(<=)(x::Py, y::Number) = pyle(x, y)
-Base.:(<)(x::Py, y::Number) = pylt(x, y)
-Base.:(>=)(x::Py, y::Number) = pyge(x, y)
-Base.:(>)(x::Py, y::Number) = pygt(x, y)
+Base.:(==)(x::Py, y::Number) = pyeq(Bool, x, y)
+Base.:(<=)(x::Py, y::Number) = pyle(Bool, x, y)
+Base.:(<)(x::Py, y::Number) = pylt(Bool, x, y)
 Base.isless(x::Py, y::Number) = pylt(Bool, x, y)
 Base.isequal(x::Py, y::Number) = pyeq(Bool, x, y)
 
-Base.:(==)(x::Number, y::Py) = pyeq(x, y)
-Base.:(!=)(x::Number, y::Py) = pyne(x, y)
-Base.:(<=)(x::Number, y::Py) = pyle(x, y)
-Base.:(<)(x::Number, y::Py) = pylt(x, y)
-Base.:(>=)(x::Number, y::Py) = pyge(x, y)
-Base.:(>)(x::Number, y::Py) = pygt(x, y)
+Base.:(==)(x::Number, y::Py) = pyeq(Bool, x, y)
+Base.:(<=)(x::Number, y::Py) = pyle(Bool, x, y)
+Base.:(<)(x::Number, y::Py) = pylt(Bool, x, y)
 Base.isless(x::Number, y::Py) = pylt(Bool, x, y)
 Base.isequal(x::Number, y::Py) = pyeq(Bool, x, y)