+++ test/runtests.jl
@@ -6,4 +6,12 @@ if haskey(ENV, "CONDA_PREFIX")
     delete!(ENV, "SSL_CERT_DIR")
 end
 
+@testmodule Setup begin
+    using PythonCall
+    # test if we are in CI
+    ci = get(ENV, "CI", "") == "true"
+    # test if we have all dev conda deps
+    devdeps = PythonCall.C.CTX.which == :CondaPkg
+end
+
 @run_package_tests